requirements.txt
fastapi==0.116.1
pdfplumber==0.11.7
psycopg[binary]==3.2.9
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.5
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
